[stable/concourse] Add optional datadog metrics config
Concourse 3.13.0 adds support for exporting metrics to Datadog (see
concourse/atc#269). This adds support for the new flags, and bumps the
Concourse version.

NOTE: This is a modification to kubernetes/charts#5650 and should be
removed when that is merged. Usage of this fork will also need to be
updated to match the changed key names.